+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.1.3
+ - added item armor types
+ - added extended frenzy button with functionality
+ - added rise at sundown button with functionality
+ - added blood heal button with functionality
+  
 ## 0.1.2
 - fix for ability roll dialog